requirements.txt
pyTelegramBotAPI==4.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10